+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [4.0.1](https://github.com/contactlab/appy/releases/tag/4.0.1)
+
+**Bug Fix:**
+
+- `[internal]` Response in ResponseError of withDecoder combinator cannot be consumed ([#360](https://github.com/contactlab/appy/issues/360))
+
+**Dependencies:**
+
+- [Security] Bump node-notifier from 8.0.0 to 8.0.1 (#347)
+
 ## [4.0.0](https://github.com/contactlab/appy/releases/tag/4.0.0)
 
 **Breaking:**